requirements.txt
# Core application dependencies
rich>=13.0.0
python-dotenv>=0.20.0 # For managing environment variables

# Data handling and scientific computing
pandas>=1.5.0
numpy>=1.23.0
scipy>=1.9.0
scikit-learn>=1.1.0 # For potential ML tasks, e.g., geocoding fallbacks
shap>=0.41.0 # For model explainability, if needed later

# Geocoding and mapping
geopy>=2.3.0
folium>=0.14.0
matplotlib>=3.7.0 # For plotting, if needed

# LLM Integration dependencies
openai>=1.3.0 # For LM Studio compatibility & official OpenAI API
tiktoken>=0.4.0 # For token counting

# LangChain and RAG dependencies
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20 # Added for components like HuggingFaceEmbeddings
langchain-huggingface>=0.0.1 # For updated HuggingFaceEmbeddings
langgraph>=0.0.20
sentence-transformers>=2.2.0 # For local embeddings
faiss-cpu>=1.7.0 # For vector store

# Development dependencies
black>=23.0.0
pylint>=2.17.0

# GUI (if applicable, assuming PyQt5 based on previous context)
# PyQt5>=5.15.0

# Other utilities
requests>=2.28.0 # General HTTP requests

# Specific versions from previous state (ensure these are covered or intentionally updated)
# torch>=2.0.0 - If still needed, should be added explicitly
# spacy>=3.5.0 - If still needed, should be added explicitly
# transformers>=4.30.0 - If still needed, should be added explicitly
